Async file loading via IoCmdList and IoCmdQueue (#12)
File IO should be nonblocking wherever possible. Using graphics APIs as inspiration, adds an async io implementation that records file IO commands to a command list and then executes them on another thread when submitted to a command queue. The number of threads started by the queue is configurable. Submitted command lists may be executed in parallel with each other if enough threads are available, but individual commands on the list will be executed in the order in which they were recorded.